extract helper funcs to genesis_config
